@@ -11,17 +11,21 @@
     "cdk": "cdk"
   },
   "devDependencies": {
-    "@aws-cdk/assert": "1.32.2",
-    "@types/jest": "^24.0.22",
-    "@types/node": "10.17.5",
-    "jest": "^24.9.0",
-    "ts-jest": "^24.1.0",
-    "aws-cdk": "1.32.2",
-    "ts-node": "^8.1.0",
-    "typescript": "~3.7.2"
+    "@aws-cdk/assert": "1.51.0",
+    "@types/jest": "^26.0.4",
+    "@types/node": "14.0.23",
+    "jest": "^26.1.0",
+    "ts-jest": "^26.1.2",
+    "aws-cdk": "1.51.0",
+    "ts-node": "^8.10.2",
+    "typescript": "~3.9.7"
   },
   "dependencies": {
-    "@aws-cdk/core": "1.32.2",
-    "source-map-support": "^0.5.16"
+    "@aws-cdk/core": "1.51.0",
+    "@aws-cdk/aws-s3": "1.51.0",
+    "@aws-cdk/aws-sns": "1.51.0",
+    "@aws-cdk/aws-sns-subscriptions": "1.51.0",
+    "@aws-cdk/aws-lambda": "1.51.0",
+    "source-map-support": "^0.5.19"
   }
 }
